+++ R/Rcpp_init.R
@@ -1,3 +1,3 @@
-#' @useDynLib dbnR
+#' @useDynLib dbnR, .registration = TRUE
 #' @importFrom Rcpp sourceCpp
 NULL