+++ tests/testthat/test-forecast_dbn.R
@@ -9,7 +9,7 @@ test_that("forecast_dbn works", {
   
   net <- learn_dbn_struc(dt_train, size = size)
   fit <- fit_dbn_params(net, f_dt_train)
-  res_fore <- forecast_ts(f_dt_test, fit, size, "pm_t_0", rep = 2,
+  res_fore <- forecast_ts(f_dt_test, fit, obj_vars = "pm_t_0", rep = 2,
                           ini = 1, len = 50, print_res = F, plot_res = F)
   
   expect_equal(1, 1)